requirements.txt
numpy
matplotlib
tqdm
tensorflow
git+https://github.com/tensorflow/datasets
Pillow